msm8996-common: Use CodeAurora ImsService implementation for RCS
There is no evidence that Qualcomm devices running a rather recent IMS
stack need to include AOSP RCS packages. RCS already worked before this
change and continues to do so, but the main goal was to get rid of some
of the RCS related log spam, which was accomplished here.

Change-Id: I219546e29d288373877e9bcd194a60ef5e09f224
Co-Authored-By: mosimchah <15335017+mosimchah@users.noreply.github.com>
Co-Authored-By: Bruno Martins <1696148+bgcngm@users.noreply.github.com>